pyproject.toml
[tool.poetry]
name = "langchain cohere"
version = "0.1.0"
description = ""
authors = ["Jackson Pacheco <jackpa1984@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.9"
langchain-cohere = "^0.3.3"
cohere = "^5.12.0"
langchain-openai = "^0.2.10"
langchain-community = "^0.3.8"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
